+++ Group Exercise 2 [Binary Tree]/bst.exe
