+++ Group Exercise 4 [Hash Table]/hash table [separate chaining].cpp
